fix(diff): circular import
